project structured and added sass
