+++ src/components/Header.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import '../styles/Header.css';
+
+function Header(props) {
+    return(
+        <header className="App-header">
+          <h1 className="App-title">Hello Buddy! :)</h1>
+        </header>
+    )
+}
+
+export default Header;
