+++ src/containers/App.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import Header from '../components/Header';
+import '../styles/App.css';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <Header/>
+      </div>
+    );
+  }
+}
+
+export default App;
