+++ src/styles/sass/Header.scss
@@ -1,0 +1,6 @@
+.App-header {
+    h1 {
+        font-size: 3rem;
+        text-align: center;
+    }
+}
